+++ vcd/resource_vcd_network_isolated.go
@@ -24,12 +24,15 @@ func resourceVcdNetworkIsolated() *schema.Resource {
 				Required: false,
 				Optional: true,
 				ForceNew: true,
+				Description: "The name of organization to use, optional if defined at provider " +
+					"level. Useful when connected as sysadmin working across different organizations",
 			},
 			"vdc": {
-				Type:     schema.TypeString,
-				Required: false,
-				Optional: true,
-				ForceNew: true,
+				Type:        schema.TypeString,
+				Required:    false,
+				Optional:    true,
+				ForceNew:    true,
+				Description: "The name of VDC to use, optional if defined at provider level",
 			},
 			"netmask": &schema.Schema{
 				Type:     schema.TypeString,